+++ app/Http/Controllers/User/UserController.php
@@ -1,45 +1,63 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\User;
 
 use App\Http\Requests\User\StoreRequest;
 use App\Models\User;
+use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
+use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Gate;
 
 class UserController extends Controller
 {
     public function index()
     {
-        $users = User::all();
+        if (!Gate::allows('user_list')) {
+            return abort(401);
+        }
+
+        $users = User::with('roles')->get();
         return view('user.index', compact('users'));
     }
 
     public function create()
     {
-        return view('user.create');
+        if (!Gate::allows('user_create')) {
+            return abort(401);
+        }
+        $roles = Role::all();
+        return view('user.create', compact('roles'));
     }
 
     public function store(StoreRequest $request)
     {
         $request->validated();
-        User::create([
+        $user = User::create([
             'name' => $request->name,
             'email' => $request->email,
             'password' =>Hash::make($request->password),
         ]);
+
+        $user->assignRole($request->role);
         return redirect()->route('users.index');
     }
 
-    public function show(string $id)
+    public function show(int $id)
     {
         //
     }
 
     public function edit(int $id)
     {
-        $user = User::where('id', $id)->first();
-        return view('user.edit', compact('user'));
+        if (!Gate::allows('user_edit')) {
+            return abort(401);
+        }
+
+        $user = User::with('roles')->where('id', $id)->first();
+        $roles = Role::all();
+        return view('user.edit', compact('user', 'roles'));
     }
 
     public function update(Request $request, int $id)
@@ -55,6 +73,12 @@ class UserController extends Controller
         $user->name = $request->name;
         $user->email = $request->email;
 
+        // Remove existing roles
+        $user->roles()->detach();
+
+        if($request->has('role')) {
+            $user->assignRole($request->role);
+        }
         // Update password if provided
         if ($request->password) {
             $request->validate(['password' => ['confirmed'],]);
@@ -70,6 +94,9 @@ class UserController extends Controller
 
     public function destroy(int $id)
     {
+        if (!Gate::allows('user_delete')) {
+            return abort(401);
+        }
         $user = User::find($id);
         return $user->delete();
     }
